Add files via upload
Write the IP address of your Kali machine in the IP section of the code and run the 'server.py' file on your computer. This will allow you to listen for the target machine. Then, convert the 'backdoor.py' file into an exe file and run it on the target machine.